+++ src/main/java/com/mateusmacial/projetogestaodepedidos/entidades/ProdutoPedido.java
@@ -12,13 +12,13 @@ import lombok.Setter;
 @Setter
 public class ProdutoPedido extends AbstractEntity {
 	private static final long serialVersionUID = 1L;
-	
-	private Integer id;
-	
+
+	private long id;
+
 	@ManyToOne
 	@JoinColumn(name = "id_pedido", referencedColumnName = "id")
 	private Pedido pedido;
-	
+
 	@ManyToOne
 	@JoinColumn(name = "id_produto", referencedColumnName = "id")
 	private Produto produto;
